+++ app/assets/stylesheets/application.scss
@@ -1,1 +1,19 @@
 @import "bootstrap";
+
+// ログイン画面
+
+.container-login {
+    @extend .container-fluid;
+    max-width: 576px;
+    padding: 2rem;
+  }
+  
+// 「ログインしました」などのフラッシュ用スタイル
+  
+  .alert-notice {
+    @extend .alert-info;
+  }
+  
+  .alert-alert {
+    @extend .alert-danger;
+  }
