--- app/assets/stylesheets/static_pages.scss
@@ -1,3 +1,0 @@
-// Place all the styles related to the StaticPages controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
