+++ app/controllers/homes_controller.rb
@@ -1,7 +1,6 @@
 class HomesController < ApplicationController
- # ログイン済ユーザーのみにアクセスを許可する
  before_action :authenticate_user!
-
+ 
  def index
   flash[:notice] = "signed_up_but_inactive" unless user_signed_in?
   @user = current_user.id
@@ -10,4 +9,11 @@ class HomesController < ApplicationController
  def new
  end
 
+ def show
+  @user = current_user
+ end
+
+ def create
+ end
+ 
 end
