+++ app/controllers/users/sessions_controller.rb
@@ -5,15 +5,9 @@ class Users::SessionsController < Devise::SessionsController
   @user = User.new
  end
  
-   #ログイン後のリダイレクト先
- def after_sign_in_path_for(resource)
-  reviews_path
- end 
-
-   #ログアウト後のリダイレクト先
- def after_sign_out_path_for(resource)
-  new_user_session_path
- end 
+ def create
+  super
+ end
 
   # protected
