+++ app/views/devise/shared/_error_messages.html.slim
@@ -1,10 +1,17 @@
-- if resource.errors.any?
+- flash.each do |msg_type, msg|
+  #alert class=("alert alert-#{msg_type}") role="alert" 
+    a.close data-dismiss="alert" href="#"  ×
+    = msg
+
+
+- if @user.errors.any?
   #error_explanation
-    h2
-      = I18n.t("errors.messages.not_saved",
-        count: resource.errors.count,
-        resource: resource.class.model_name.human.downcase)
-    ul
-      - resource.errors.full_messages.each do |message|
-        li
-          = message
+   h2
+     = I18n.t("errors.messages.not_saved",
+     count: @user.errors.count,
+     resource: @user.class.model_name.human.downcase)
+
+   ul
+     - @user.errors.full_messages.each do |message|
+       li
+         = message
