+++ app/views/devise/shared/_header.html.slim
@@ -1,11 +1,8 @@
- .app-title.navbar.navbar-expand-md.navbar-light.bg-ligth.navbar-brand BookReviewApp
-  .container
-   - if flash.notice.present?
-    .alert.alert-success= flash.notice
-
- - if user_signed_in?
-  = link_to "Account update", edit_user_registration_path
-  = link_to "Log out", destroy_user_session_path, method: :delete
- - else
-  = link_to "Sign in", new_user_registration_path
-  = link_to "Log in", new_user_session_path
+- if user_signed_in?
+/ ユーザーがログインしている場合に実行する処理
+ = link_to "Account update", edit_user_registration_path
+ = link_to "Log out", destroy_user_session_path, method: :delete
+- else
+/ ユーザーがログアウトしている場合に実行する処理
+ = link_to "Sign in", new_user_registration_path
+ = link_to "Log in", new_user_session_path
