+++ app/views/homes/index.html.slim
@@ -1,1 +1,1 @@
-= render "shared/error_messages"
+= render 'devise/shared/error_messages'
