+++ app/views/homes/show.html.slim
@@ -1,0 +1,2 @@
+h1 Homes#show
+p Find me in app/views/homes/show.html.slim
