+++ config/routes.rb
@@ -1,5 +1,7 @@
 Rails.application.routes.draw do
+  
  root to: 'homes#index'
+ get 'homes/show'
  devise_for :users, controllers: {
   registrations: 'users/registrations',
   sessions: "users/sessions",
