--- spec/controllers/static_pages_controller_spec.rb
@@ -1,19 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe StaticPagesController, type: :controller do
-
-  describe "GET #top" do
-    it "returns http success" do
-      get :top
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET #home" do
-    it "returns http success" do
-      get :home
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
