+++ spec/views/homes/show.html.slim_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "homes/show.html.slim", type: :view do
+  pending "add some examples to (or delete) #{__FILE__}"
+end
